+++ biohansel/create/cluster_generator.py
@@ -233,5 +233,4 @@ def df_to_subtypes_dict(cluster_genomes_df: pd.DataFrame) -> Dict[str, str]:
     final_cluster_dict = {}
     for genome, row in cluster_genomes_df.apply(row_subtype, axis=1).iteritems():
         final_cluster_dict[genome] = row
-
     return final_cluster_dict
